[Refactor]#141 presentation.network 패키지를 data패키지로 이동
 - clean architecture에 따르면, Database, Api관련 코드는 data layer에 위치
